fwprovider: Move data source/ephemeral resource/resource wrappers to 'wrap.go'.
